WIP fixing thread sleep times, working on udp/tcp comm tests
