+++ src/communications/communicator/constants.py
@@ -1,2 +1,3 @@
+THREAD_SLEEP_TIME = 0.005 # in seconds
 DEFAULT_SERVER_ADDRESS = ('127.0.0.1', 7777)
 DEFAULT_GAME_SERVER_ADDRESS = ('127.0.0.1', 7778)
