+++ src/communications/communicator/receiver.py
@@ -2,11 +2,7 @@ import socket
 import time
 from threading import Thread
 
-# TODO: temporary, clean this up later
-SLEEP_TIME = 50
-
-# TODO: implement decode(message) method
-# TODO: implement DecodeError if message cannot be decoded successfully
+from src.communications.communicator.constants import THREAD_SLEEP_TIME
 
 
 class Receiver(Thread):
@@ -26,7 +22,7 @@ class Receiver(Thread):
                 message = self._receiveMessage()
                 self.communicator.enqueueTask(message)
             except socket.error:
-                time.sleep(SLEEP_TIME)
+                time.sleep(THREAD_SLEEP_TIME)
             except DecodeError:
                 # TODO: may want to log whenever this happens
                 pass
